serverside/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
numpy==2.0.1
pandas==2.2.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stripe==10.8.0
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
